Fix crash when working with ECR registry
This was one of the hardest things I have ever had to track down.
Node.js projects truely are the worst, and I will never willingly work
with them.

AWS ECR, when GETing a blob, will 307 you to the actual backing bucket
URL where it is stored. Since this is on a different domain, ECR sends a
short-lived auth token as part of the URL.

If you then send an Authorization header as part of the second GET, ECR
explodes. I have no idea if this is valid under the Docker Hub API v2,
but my guess is that the spec is not specific enough to define behaviour
here.

fetch-h2.js, the library used by the Docker library used by this action,
will send any headers of the first request to the second. And the Docker
library uses a custom header to set the auth instead of the correct auth
mechanisms in the API.

The "right" fix is to have the Docker library use the right auth API
points and set `credentials: "same-origin"`. However, the fetch-h2
library has a TODO that that mechanism is not implemented yet.

Meaning this hacked fix will have to do.

fetch-h2.js is basically abandonware, so even if I did the work to make
a real fix and make a PR, it would not be merged. I don't want to
maintain a fork of fetch-h2.js, so I am keeping patches locally. They
are automatically applied with build.